+++ envsetup.sh
@@ -217,5 +217,5 @@ siot_goreleaser_build() {
 siot_goreleaser_release() {
   #TODO add depend build to goreleaser config
   siot_build_dependencies --optimize
-  goreleaser release
+  goreleaser release --rm-dist
 }